from uint8 to float32
float32 more appropriate for Tensorflow GPU

+++ model/feeder.py
@@ -20,12 +20,12 @@ class ImageFeeder(tf.keras.utils.Sequence):
         i = idx * self.batch_size
         batch_input_data = self.features_dataset[i : i + self.batch_size]
         batch_labels_data = self.labels_dataset[i : i + self.batch_size]
-        x = np.zeros((self.batch_size,) + self.img_size + (1,), dtype="uint8")
+        x = np.zeros((self.batch_size,) + self.img_size + (1,), dtype="float32")
         for j in range(len(batch_input_data)):
             img = batch_input_data[j]
             img = np.expand_dims(img, axis=2)
             x[j] = img
-        y = np.zeros((self.batch_size,) + self.img_size + (1,), dtype="uint8")
+        y = np.zeros((self.batch_size,) + self.img_size + (1,), dtype="float32")
         for j in range(len(batch_labels_data)):
             label = batch_labels_data[j]
             label = np.expand_dims(label, axis=2)
